refactor: Change file path of analyzer
